Replaced unseen docstrings with comments.
Python generally interprets the double-underscore prefix of method names
such that the docstrings associated with those methods are not displayed
(there are exceptions).  In addition, the whole point of
double-underscoring a name is to keep the method private.

So...  I went through and replaced private-method docstrings with
comments that contain the same content.