HexDump.py; Moved some code to a different file.
The code that is now in carnaval/common/HexDump.py was originally in
nbt/NBT_Core.py, but those functions are also quite useful in other
subpackages.  Unfortunately, several attempts to import the functions
from one subpackage to another have failed.  The mechanisms outlined
in the documentation all seem to result in exceptions.  The working
solution is to create a soft-link to the module--a kludge at best.