SMB_Status: Rewrite; NTStatus inherits from long.
The NTStatus class is now a descendant of the built-in long class.  It behaves like a long ionteger, but has added attributes and methods.

In addition to storing the error code value, we also store the error code name (as a string) and error message.  These attributes are available as properties.

A set of SMB error codes are instanciated globally, and made available when the modue is imported.  There is also a class-level dictionary that is used to look up instances by error code.